Adds Role model, fixes Production model references
This adds a Role model which belongs_to Production. As
I was adding this I noticed an error in the Production model,
which had a `belongs_to :user` where it should have had
a `belongs_to :show` - this fixes that reference and makes
relevant changes to tests and factories.

Note that I am not sure about the `belongs_to :person` in the
Role model. It would mean that you couldn't for instance
have two actors play the same role on different nights, or
assign multiple actors to a chorus or something like that,
so this might need to be revisited in the future.