Major changes to data model
After talking with Patsy, who is a subject matter expert, we decided
that this project should start with a way to crowdsource show information
that any user could access. In order to do this some of the associations
needed to be changed.

I also took this opportunity to add an Author model