Adds Author views and constants
This adds a suite of views for the `Author` model, and makes a few
changes as well. Notably this adds in the missing :lgbt attribute to
the Author model, and adds a service class to hold constants like the
hashes for gender_identity and ethnicity which will be used in multiple
places.

As a sidenote, it makes me uncomfortable to include these values, but
in talking with Patsy it became clear that this was quite important. We
want, for instance, to allow a theater to find plays written by women
or people of color. Similarly we want to be able to allow directors the
ability to select a diverse cast when looking for actors for their
 production. It may be naive, but my hope is that this feature will
 be used for good rather than for evil.