Layout changes
I wanted to have a right-hand details panel where forms and other data could be displayed. This a step in that direction where we have most of the views using the new layout and the hotwire pattern. I also removed alpinejs which I wasn't using. I am starting to think that stimulus and turbo might be sufficient for my js needs, at least until I start working with the stage view.