Schema and model changes to show, author, and role
I found a database of shows, roles, and authors which I can use to prepopulate our database with useful information. In order to get the associations right, I needed to add show_code and author_code columns to their respective models. I also removed the relationship between roles and people, which was not correct. We'll need to revisit that at some point, but this is not the time.