Makes author<>show relationship many-to-many, associated UI changes
Fundamentally this is about making the relationship between authors and shows many-to-many, but as part of that we needed to make some meaningful changes to the form ui so that multiple authors could be selected. In order to do that, I brought back jquery so that we could use the select2 library which works really well for multi-selects with typeahead.