Switches out a lot of the template styles for daisyui
I'm switching over to daisyui which I like a lot. Things
are starting to look pretty good, though there are probably
some forms and other areas that I've missed, so this
will be a work-in-progress, but it's a pretty good start.

Not sure I like the primary color, might need to work on that