Fixes datatable initialization, fixes contact form spec
Datatables were not getting properly destroyed, which meant that if the user navigated forward/back the datatable would keep getting initialized and cause all kinds of awful errors. This will destroy the stale datatable each time and make a fresh one. There was also a bug in the contact form spec which I fixed while I was at it.