Permit blank first and last names if the other is present.
We want every author to have a name, but we don't want to exclude first-name-only people (Cher) or last-name-only people (Euripides) so this adds logic to the presence validations to allow blank if the other is present. I've gotten feedback that a custom validator might be easier to understand so this maybe should be revisited.