Removes datatables and name_of_person
This removes some dependencies. Most importantly I am removing datatables and replacing its functionality with a combination of kaminari pagination and advanced search via ransack. I am doing this because there were some performance issues with server-side datatables and there was a FOUC when initializing the datatables on each page as well. This solution loses the clickable headers, which may need to be replaced by links to server-side sorting.