Links author and show link buttons to correct page of paginated results
There was a problem where the author and show links were not pointing to the correct page of results on the `#show` view, so I had to find a solution for getting the page a record is on so that it can be linked correctly. Shows were a little tricky because of the join, but I think I got it right.