Adds act/scene/french_scene models
This adds models for Act, Scene, and French Scene. These are all more or
less the same at the moment, and it's possible that they will stay
that way, but it seemed like it would be simpler and clearer to keep
them as separate models rather than having weird nested associations.