Add collapsible elements for Roles and Acts/Scenes
The Role and Act/Scene lists were starting to get unwieldy on the show form, so I added
collapsible elements surrounding them so that they could be shown and hidden as-needed.
I'm using the daisyui collapsible component, which I think will probably be fine, but
it is a little less functional than a javascript-based collapse, so I will probably
need to keep an eye on it and get user feedback.