Improve XLSX export, add button
This changes the format of the xlsx import so that the acts/scenes are the x-axis and the roles are the y-axis. It also moves the export logic to a service object. I don't have great tests yet, which is weird because simplecov seems to think I still have 100% coverage, but I am pretty sure I don't. I plan to at least add coverage for the service object and the controller action.